Fix previous commit.

+++ src/main/java/com/gregb/library/androidhelper/helper/Convertor.java
@@ -17,7 +17,7 @@ public class Convertor {
   public static long toPrimitive(Long l){
     return Validator.isValid(l) ? l : 0;
   }
-  public static long toPrimitive(Float f){
+  public static float toPrimitive(Float f){
     return Validator.isValid(f) ? f : 0;
   }
